+++ src/internal/during.js
@@ -7,11 +7,17 @@ import { msElapsed } from './msElapsed'
 
 const durationTypeErrorMessage =
   'second argument (duration) of during should be a number'
+const durationRangeErrorMessage =
+  'second argument (duration) of during should be a positive number'
 
 const during =
   (scheduler) => (duration) => {
     if (typeof duration !== 'number') {
       throw new TypeError(durationTypeErrorMessage)
+    }
+
+    if (duration <= 0) {
+      throw new RangeError(durationRangeErrorMessage)
     }
 
     return msElapsed(scheduler)